+++ igneous/tasks/__init__.py
@@ -1,4 +1,5 @@
 from .skeletonization import SkeletonTask, UnshardedSkeletonMergeTask, ShardedSkeletonMergeTask
+from .mesh import MeshTask, MeshManifestTask, GrapheneMeshTask
 from .tasks import (
   IngestTask, HyperSquareConsensusTask, 
   MeshTask, MeshManifestTask, DownsampleTask, QuantizeTask,
