+++ igneous/tasks/tasks.py
@@ -225,389 +225,6 @@ class QuantizeTask(RegisteredTask):
     destvol = CloudVolume(self.dest_layer_path, mip=self.mip)
     downsample_and_upload(image, bounds, destvol, self.shape, mip=self.mip, axis='z')
 
-
-class MeshTask(RegisteredTask):
-  def __init__(self, shape, offset, layer_path, **kwargs):
-    """
-    Convert all labels in the specified bounding box into meshes
-    via marching cubes and quadratic edge collapse (github.com/seung-lab/zmesh).
-
-    Required:
-      shape: (sx,sy,sz) size of task
-      offset: (x,y,z) offset from (0,0,0)
-      layer_path: neuroglancer/cloudvolume dataset path
-
-    Optional:
-      lod: (uint) level of detail to record these meshes at
-      mip: (uint) level of the resolution pyramid to download segmentation from
-      simplification_factor: (uint) try to reduce the number of triangles in the 
-        mesh by this factor (but constrained by max_simplification_error)
-      max_simplification_error: The maximum physical distance that
-        simplification is allowed to move a triangle vertex by. 
-      mesh_dir: which subdirectory to write the meshes to (overrides info file location)
-      remap_table: agglomerate segmentation before meshing using { orig_id: new_id }
-      generate_manifests: (bool) if it is known that the meshes generated by this 
-        task will not be cropped by the bounding box, avoid needing to run a seperate
-        MeshManifestTask pass by generating manifests on the spot.
-
-      These two options are used to allow sufficient overlap for trivial mesh stitching
-      between adjacent tasks.
-
-        low_padding: (uint) expand the bounding box by this many pixels by subtracting
-          this padding from the minimum point of the bounding box on all axes.
-        high_padding: (uint) expand the bounding box by this many pixels adding
-          this padding to the maximum point of the bounding box on all axes.
-
-      parallel_download: (uint: 1) number of processes to use during the segmentation download
-      cache_control: (str: None) specify the cache-control header when uploading mesh files
-      dust_threshold: (uint: None) don't bother meshing labels strictly smaller than this number of voxels.
-      encoding: (str) 'precomputed' (default) or 'draco'
-      draco_compression_level: (uint: 1) only applies to draco encoding
-      draco_create_metadata: (bool: False) only applies to draco encoding
-      progress: (bool: False) show progress bars for meshing 
-      object_ids: (list of ints) if specified, only mesh these ids
-      fill_missing: (bool: False) replace missing segmentation files with zeros instead of erroring
-      spatial_index: (bool: False) generate a JSON spatial index of which meshes are available in
-        a given bounding box. 
-      sharded: (bool: False) If True, upload all meshes together as a single pickled 
-        fragment file. 
-      timestamp: (int: None) (graphene only) use the segmentation existing at this
-        UNIX timestamp.
-    """
-    super(MeshTask, self).__init__(shape, offset, layer_path, **kwargs)
-    self.shape = Vec(*shape)
-    self.offset = Vec(*offset)
-    self.layer_path = layer_path
-    self.options = {
-      'cache_control': kwargs.get('cache_control', None),
-      'draco_compression_level': kwargs.get('draco_compression_level', 1),
-      'draco_create_metadata': kwargs.get('draco_create_metadata', False),
-      'dust_threshold': kwargs.get('dust_threshold', None),
-      'encoding': kwargs.get('encoding', 'precomputed'),
-      'fill_missing': kwargs.get('fill_missing', False),
-      'generate_manifests': kwargs.get('generate_manifests', False),
-      'high_padding': kwargs.get('high_padding', 1),
-      'low_padding': kwargs.get('low_padding', 0),
-      'lod': kwargs.get('lod', 0),
-      'max_simplification_error': kwargs.get('max_simplification_error', 40),
-      'simplification_factor': kwargs.get('simplification_factor', 100),
-      'mesh_dir': kwargs.get('mesh_dir', None),
-      'mip': kwargs.get('mip', 0),
-      'object_ids': kwargs.get('object_ids', None),
-      'parallel_download': kwargs.get('parallel_download', 1),
-      'progress': kwargs.get('progress', False),
-      'remap_table': kwargs.get('remap_table', None),
-      'spatial_index': kwargs.get('spatial_index', False),
-      'sharded': kwargs.get('sharded', False),
-      'timestamp': kwargs.get('timestamp', None),
-      'agglomerate': kwargs.get('agglomerate', True),
-      'stop_layer': kwargs.get('stop_layer', 2),
-    }
-    supported_encodings = ['precomputed', 'draco']
-    if not self.options['encoding'] in supported_encodings:
-      raise ValueError('Encoding {} is not supported. Options: {}'.format(
-        self.options['encoding'], ', '.join(supported_encodings)
-      ))
-    self._encoding_to_compression_dict = {
-      'precomputed': True,
-      'draco': False,
-    }
-
-  def execute(self):
-    self._volume = CloudVolume(
-      self.layer_path, self.options['mip'], bounded=False,
-      parallel=self.options['parallel_download'], 
-      fill_missing=self.options['fill_missing']
-    )
-    self._bounds = Bbox(self.offset, self.shape + self.offset)
-    self._bounds = Bbox.clamp(self._bounds, self._volume.bounds)
-
-    self.progress = bool(self.options['progress'])
-
-    self._mesher = zmesh.Mesher(self._volume.resolution)
-
-    # Marching cubes loves its 1vx overlaps.
-    # This avoids lines appearing between
-    # adjacent chunks.
-    data_bounds = self._bounds.clone()
-    data_bounds.minpt -= self.options['low_padding']
-    data_bounds.maxpt += self.options['high_padding']
-
-    self._mesh_dir = self.get_mesh_dir()
-
-    if self.options['encoding'] == 'draco':
-      self.draco_encoding_settings = self._compute_draco_encoding_settings()
-
-    # chunk_position includes the overlap specified by low_padding/high_padding
-    # agglomerate, timestamp, stop_layer only applies to graphene volumes, 
-    # no-op for precomputed
-    data = self.volume.download(
-      data_bounds, 
-      agglomerate=self.options['agglomerate'], 
-      timestamp=self.options['timestamp'], 
-      stop_layer=self.options['stop_layer']
-    )
-
-    if not np.any(data):
-      return
-
-    data = self._remove_dust(data, self.options['dust_threshold'])
-    data = self._remap(data)
-
-    if self.options['object_ids']:
-      data = fastremap.mask_except(data, self.options['object_ids'], in_place=True)
-
-    data, renumbermap = fastremap.renumber(data, in_place=True)
-    renumbermap = { v:k for k,v in renumbermap.items() }
-    self._compute_meshes(data, renumbermap)
-
-  def get_mesh_dir(self):
-    if self.options['mesh_dir'] is not None:
-      return self.options['mesh_dir']
-    elif 'mesh' in self._volume.info:
-      return self._volume.info['mesh']
-    else:
-      raise ValueError("The mesh destination is not present in the info file.")
-
-  def _compute_draco_encoding_settings(self):
-    def calculate_quantization_bits_and_range(min_quantization_range, max_draco_bin_size, draco_quantization_bits=None):
-        if draco_quantization_bits is None:
-            draco_quantization_bits = np.ceil(
-                np.log2(min_quantization_range / max_draco_bin_size + 1))
-        num_draco_bins = 2 ** draco_quantization_bits - 1
-        draco_bin_size = np.ceil(min_quantization_range / num_draco_bins)
-        draco_quantization_range = draco_bin_size * num_draco_bins
-        if draco_quantization_range < min_quantization_range + draco_bin_size:
-            if draco_bin_size == max_draco_bin_size:
-                return calculate_quantization_bits_and_range(min_quantization_range, max_draco_bin_size, draco_quantization_bits + 1)
-            else:
-                draco_bin_size = draco_bin_size + 1
-                draco_quantization_range = draco_quantization_range + num_draco_bins
-        return draco_quantization_bits, draco_quantization_range, draco_bin_size
-
-    min_quantization_range = max((self.shape + self.options['low_padding'] + self.options['high_padding']) * self._volume.resolution)
-    max_draco_bin_size = np.floor(min(self._volume.resolution) / np.sqrt(2))
-    draco_quantization_bits, draco_quantization_range, draco_bin_size = calculate_quantization_bits_and_range(
-        min_quantization_range, max_draco_bin_size)
-    draco_quantization_origin = self.offset - (self.offset % draco_bin_size)
-    return {
-        'quantization_bits': draco_quantization_bits,
-        'compression_level': self.options['draco_compression_level'],
-        'quantization_range': draco_quantization_range,
-        'quantization_origin': draco_quantization_origin,
-        'create_metadata': self.options['draco_create_metadata']
-    }
-
-  def _remove_dust(self, data, dust_threshold):
-    if dust_threshold:
-      segids, pxct = fastremap.unique(data, return_counts=True)
-      dust_segids = [ sid for sid, ct in zip(segids, pxct) if ct < int(dust_threshold) ]
-      data = fastremap.mask(data, dust_segids, in_place=True)
-
-    return data
-
-  def _remap(self, data):
-    if self.options['remap_table'] is None:
-      return data 
-
-    self.options['remap_table'] = {
-      int(k): int(v) for k, v in self.options['remap_table'].items()
-    }
-
-    remap = self.options['remap_table']
-    remap[0] = 0
-
-    data = fastremap.mask_except(data, list(remap.keys()), in_place=True)
-    return fastremap.remap(data, remap, in_place=True)
-
-  def _compute_meshes(self, data, renumbermap):
-    data = data[:, :, :, 0].T
-    self._mesher.mesh(data)
-    del data
-
-    bounding_boxes = {}
-    meshes = {}
-
-    for obj_id in tqdm(self._mesher.ids(), disable=(not self.progress), desc="Mesh"):
-      remapped_id = renumbermap[obj_id]
-      mesh_binary, mesh_bounds = self._create_mesh(obj_id)
-      bounding_boxes[remapped_id] = mesh_bounds.to_list()
-      meshes[remapped_id] = mesh_binary
-
-    if self.options['sharded']:
-      self._upload_batch(meshes, self._bounds)
-    else:
-      self._upload_individuals(meshes, self.options['generate_manifests'])
-
-    if self.options['spatial_index']:
-      self._upload_spatial_index(self._bounds, bounding_boxes)
-
-  def _upload_batch(self, meshes, bbox):
-    with SimpleStorage(self.layer_path, progress=self.options['progress']) as stor:
-      # Create mesh batch for postprocessing later
-      stor.put_file(
-        file_path="{}/{}.frags".format(self._mesh_dir, bbox.to_filename()),
-        content=pickle.dumps(meshes),
-        compress='gzip',
-        content_type="application/python-pickle",
-        cache_control=False,
-      )
-
-  def _upload_individuals(self, mesh_binaries, generate_manifests):
-    with Storage(self.layer_path) as storage:
-      for segid, mesh_binary in mesh_binaries.items():
-        storage.put_file(
-          file_path='{}/{}:{}:{}'.format(
-            self._mesh_dir, segid, self.options['lod'],
-            self._bounds.to_filename()
-          ),
-          content=mesh_binary,
-          compress=self._encoding_to_compression_dict[self.options['encoding']],
-          cache_control=self.options['cache_control']
-        )
-
-        if generate_manifests:
-          fragments = []
-          fragments.append('{}:{}:{}'.format(
-            segid, self.options['lod'],
-            self._bounds.to_filename())
-          )
-
-          storage.put_file(
-            file_path='{}/{}:{}'.format(
-              self._mesh_dir, segid, self.options['lod']
-            ),
-            content=json.dumps({"fragments": fragments}),
-            content_type='application/json',
-            cache_control=self.options['cache_control']
-          )
-
-
-  def _create_mesh(self, obj_id):
-    mesh = self._mesher.get_mesh(
-      obj_id,
-      simplification_factor=self.options['simplification_factor'],
-      max_simplification_error=self.options['max_simplification_error']
-    )
-
-    self._mesher.erase(obj_id)
-
-    resolution = self._volume.resolution
-    offset = self._bounds.minpt - self.options['low_padding']
-    mesh.vertices[:] += offset * resolution
-
-    mesh_bounds = Bbox(
-      np.amin(mesh.vertices, axis=0), 
-      np.amax(mesh.vertices, axis=0)
-    )
-
-    if self.options['encoding'] == 'draco':
-      mesh_binary = DracoPy.encode_mesh_to_buffer(
-        mesh.vertices.flatten('C'), mesh.faces.flatten('C'), 
-        **self.draco_encoding_settings
-      )
-    elif self.options['encoding'] == 'precomputed':
-      mesh_binary = mesh.to_precomputed()
-
-    return mesh_binary, mesh_bounds
-
-  def _upload_spatial_index(self, bbox, mesh_bboxes):
-    with SimpleStorage(self.layer_path, progress=self.options['progress']) as stor:
-      stor.put_file(
-        file_path="{}/{}.spatial".format(self._mesh_dir, bbox.to_filename()),
-        content=jsonify(mesh_bboxes).encode('utf8'),
-        compress='gzip',
-        content_type="application/json",
-        cache_control=False,
-      )
-
-class GrapheneMeshTask(MeshTask):
-  def __init__(self, *args, **kwargs):
-    kwargs['encoding'] = 'draco'
-    kwargs['sharded'] = True
-    kwargs['spatial_index'] = False # use PyChunkGraph for this info instead
-    kwargs['draco_compression_level'] = 1
-    kwargs['draco_create_metadata'] = True
-    super(GrapheneMeshTask, MeshTask).__init__(*args, **kwargs)
-
-  def upload_meshes(self, meshes, bounding_boxes):
-    layer_id = 2 # we will always mesh layer 2
-    reader = self._volume.mesh.readers[layer_id] 
-    spec = reader.spec 
-
-    shard_binary = spec.synthesize_shard(meshes)
-    shard_filename = reader.get_filename(list(meshes.keys())[0])
-
-    with SimpleStorage(self._volume.cloudpath) as stor:
-      stor.put_file(
-        file_path="{}/initial/{}/{}".format(self._mesh_dir, layer_id, shard_filename),
-        content=shard_binary,
-        compress=None,
-        content_type="application/octet-stream",
-        cache_control="no-cache",
-      )
-
-class MeshManifestTask(RegisteredTask):
-  """
-  Finalize mesh generation by post-processing chunk fragment
-  lists into mesh fragment manifests.
-  These are necessary for neuroglancer to know which mesh
-  fragments to download for a given segid.
-
-  If we parallelize using prefixes single digit prefixes ['0','1',..'9'] all meshes will
-  be correctly processed. But if we do ['10','11',..'99'] meshes from [0,9] won't get
-  processed and need to be handle specifically by creating tasks that will process
-  a single mesh ['0:','1:',..'9:']
-  """
-
-  def __init__(self, layer_path, prefix, lod=0, mesh_dir=None):
-    super(MeshManifestTask, self).__init__(layer_path, prefix)
-    self.layer_path = layer_path
-    self.lod = lod
-    self.prefix = prefix
-    self.mesh_dir = mesh_dir
-
-  def execute(self):
-    with Storage(self.layer_path) as storage:
-      self._info = json.loads(storage.get_file('info').decode('utf8'))
-
-      if self.mesh_dir is None and 'mesh' in self._info:
-        self.mesh_dir = self._info['mesh']
-
-      self._generate_manifests(storage)
-
-  def _get_mesh_filenames_subset(self, storage):
-    prefix = '{}/{}'.format(self.mesh_dir, self.prefix)
-    segids = defaultdict(list)
-
-    for filename in storage.list_files(prefix=prefix):
-      filename = os.path.basename(filename)
-      # `match` implies the beginning (^). `search` matches whole string
-      matches = re.search(r'(\d+):(\d+):', filename)
-
-      if not matches:
-        continue
-
-      segid, lod = matches.groups()
-      segid, lod = int(segid), int(lod)
-
-      if lod != self.lod:
-        continue
-
-      segids[segid].append(filename)
-
-    return segids
-
-  def _generate_manifests(self, storage):
-    segids = self._get_mesh_filenames_subset(storage)
-    for segid, frags in segids.items():
-      storage.put_file(
-          file_path='{}/{}:{}'.format(self.mesh_dir, segid, self.lod),
-          content=json.dumps({"fragments": frags}),
-          content_type='application/json',
-      )
-
-
 # class BigArrayTask(RegisteredTask):
 #   def __init__(self, layer_path, chunk_path, chunk_encoding, version):
 #     super(BigArrayTask, self).__init__(
